tali/requirements.txt
djangorestframework==3.15.1
six==1.16.0
django-registration==3.4
pika==1.3.2
celery==5.4.0